refactor(docx): lift the line-layout engine out of renderer.ts (verbatim, B2 phase boundary)
The B2 text-layout unification (paragraph #684/#689, table #693, textbox #697)
grew a cohesive layout-engine phase inside renderer.ts: the pure segmentation +
line-breaking + measurement kernel that BOTH the paginator and the paint pass
call to turn a paragraph's runs into laid-out lines and line-box heights. This
lifts that phase verbatim into a new packages/docx/src/line-layout.ts, following
the #635/#636 method (byte-identical function bodies, no behaviour change).

Moved (verbatim): the kernel (buildSegments / layoutLines / rescaleLayoutLines),
the layout data types (LayoutSeg family, LayoutLine, WrapLayoutCtx, DocGridCtx,
MathRender + the mathRenders extent cache), the segmentation helpers (splitBy*,
splitDigitGroups, fitCJKPrefix, hasCJKBreakOpportunity, isRtlBidiLang, …), the
tab-stop math (nextTabStop / resolveDefaultTabPt / DEFAULT_TAB_PT), the font
resolution (buildFont / normalizeFontFamily + fallback-chain builders), and the
metric/grid math (lineBoxHeight, correctedLineMetrics, gridWidth family,
paragraphMarkLineHeight, EAST_ASIAN_RE, …). Paint (drawParagraphLine,
renderShapeText), pagination, tables, floats and anchors stay in renderer.ts.

The split is one-directional at runtime: renderer.ts → line-layout.ts. The only
back-reference is the RenderState / DecodedImage TYPE (import type, erased) —
the same idiom frame-geometry.ts / anchor-geometry.ts use — so there is no
import cycle and no re-export layer. renderer.ts now imports the kernel it
drives; the previously public normalizeFontFamily / splitDigitGroups /
lineBoxHeight test imports point at ./line-layout.js directly.

renderer.ts: 9924 → 7789 lines. All 520 docx tests pass, incl. the
characterization / reuse-identity suites that fix the render-command stream.
